# gen_openapi/swallow_channel_from_service/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /v1/product:
        post:
            tags:
                - ServiceSwallow
            operationId: ServiceSwallow_UpsertProduct
            parameters:
                - name: id
                  in: query
                  description: 'item ID, if filed exist: Update else Create'
                  schema:
                    type: integer
                    format: int32
                - name: jsonMetadata
                  in: query
                  description: |-
                    client side, data, now we not using them,
                     but in future if field will participate to internal logic need to be added to Product fields
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/swallow_channel_to_service.Product'
    /v1/product/{id}:
        get:
            tags:
                - ServiceSwallow
            description: Get on product by some field
            operationId: ServiceSwallow_GetProduct
            parameters:
                - name: id
                  in: path
                  description: Filter documents by ID Field
                  required: true
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/swallow_channel_to_service.Product'
...
components:
    schemas:
...
        swallow_channel_to_service.Product:
            type: object
            properties:
                id:
                    type: integer
                    description: item ID
                    format: int32
                jsonMetadata:
                    type: string
                    description: |-
                        client side, data, now we not using them,
                         but in future if field will participate to internal logic need to be added to Product fields
            description: Can be Ticket/Reservation
...